requirements.txt
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
autopep8==2.3.1
billiard==4.2.0
black==24.10.0
celery==5.4.0
certifi==2023.11.17
cffi==1.16.0
channels==4.2.0
channels-redis==4.2.0
charset-normalizer==3.3.2
chardet==5.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.39.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==42.0.3
daphne==4.1.2
Django==5.0.8
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-channels==0.7.0
django-cors-headers==4.3.1
django-debug-toolbar==4.4.6
django-jazzmin==3.0.1
django-ninja==1.2.2
django-phonenumber-field==8.0.0
django-redis==5.4.0
django-rest-framework==0.1.0
django-restframework==0.0.1
django-silk==5.3.1
django-timezone-field==7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
docutils==0.21.2
drf-yasg==1.21.7
email_validator==2.2.0
exceptiongroup==1.2.1
filetype==1.2.0
flake8==7.1.1
gprof2dot==2024.6.6
greenlet==3.1.1
gunicorn==21.2.0
honcho==2.0.0
honeybadger==0.19.0
humanize==4.11.0
hyperlink==21.0.0
idna==3.6
importlib_metadata==8.5.0
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
keyring==25.4.1
kombu==5.4.2
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.5.0
msgpack==1.0.8
mypy-extensions==1.0.0
mysqlclient==2.2.1
nh3==0.2.18
oauthlib==3.2.2
olefile==0.47
packaging==23.2
pathspec==0.12.1
phonenumbers==8.13.43
pillow==10.2.0
pkginfo==1.10.0
platformdirs==4.3.6
pluggy==1.5.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.12.1
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.0.0
pytest==8.2.1
pytest-django==4.8.0
pytest-mock==3.14.0
python-crontab==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
readme_renderer==44.0
redis==5.0.1
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.2
SecretStorage==3.3.3
service-identity==24.1.0
setuptools==75.8.0
six==1.16.0
SQLAlchemy==2.0.37
sqlparse==0.4.4
twine==5.1.1
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
whitenoise==6.8.2
zipp==3.20.2
zope.interface==6.2